requirements.txt
Flask==2.1.3
pandas>=1.5
mysql-connector-python==8.0.31
sentence-transformers==2.2.2
scikit-learn==1.1.3
nltk==3.7
gunicorn==20.1.0
Werkzeug==2.1.2
numpy>=1.24